[기능 구현] 김진도 - 학생 퀴즈방 입장 , 퀴즈 문제 고르기

+++ src/main/java/com/samsamgyeesam/studyingvally/domain/quiz/dto/QuizDTO.java
@@ -8,7 +8,7 @@ import lombok.*;
 @Setter
 @ToString
 public class QuizDTO {
-    private int quizNo;
+    private Long quizNo;
     private String quizTitle;
-    private int chapNo; // 파라미터로 받을 챕터 번호
+    private Long chapNo; // 파라미터로 받을 챕터 번호
 }
